{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["Erick Navarro <erick@navarro.io>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
psycopg2-binary = "^2.8.5"
dj-database-url = "^0.5.0"
django-model-utils = "^4.0.0"
gunicorn = "^20.0.4"
raven = "^6.10.0"
django-crispy-forms = "^1.9.1"
pillow = "^7.1.2"
celery = "^4.4.5"
redis = "^3.5.3"
django-celery-results = "^1.2.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pep8-naming = "^0.11.1"
black = "^19.10b0"
pytest = "^5.4.3"
pytest-django = "^3.9.0"
pytest-coverage = "^0.0"
mixer = "^6.1.3"

[tool.flake8]
exclude = ".git,*migrations*,assets,node_modules"
max-line-length = 119

[tool.isort]
line_length = 119

[tool.pytest]
norecursedirs = ".git assets node_modules"

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | assets
  | node_modules
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
